backend/requirements.txt
# Web框架
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# 阿里云SDK
dashscope==1.14.1

# 向量数据库和检索
chromadb==0.4.22
sentence-transformers==2.3.1

# 数据处理和验证
pydantic==2.5.3
pydantic-settings==2.1.0

# 异步HTTP客户端
httpx==0.26.0
aiohttp==3.9.1

# 图片处理
Pillow==10.2.0

# 数据库（可选，用于元数据存储）
sqlalchemy==2.0.25
aiosqlite==0.19.0

# 工具库
python-dotenv==1.0.0
tenacity==8.2.3

# 日志和监控
loguru==0.7.2

# 测试
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# 代码质量
black==24.1.1
mypy==1.8.0

